+++ src/app/projects/project-list/project-list-routing.module.ts
@@ -8,7 +8,8 @@ const routes: Routes = [
     component: ProjectListComponent,
     data: {
       title: 'List View of Projects',
-      focush1: true
+      focush1: true,
+      description: 'Browse all land and water planning projects in British Columbia. View active planning engagements in list format.'
     }
   }
 ];